Functions for fixedPrices and description in README.md
